seperate test for setpackage
